requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
itsdangerous==2.1.2
gunicorn==21.2.0
